refactor(trusty-common): split dashboard into submodules under 500-SLOC cap (#607)
Replace the 745-SLOC monolith `monitor/dashboard.rs` with a four-file module:
  - `dashboard/types.rs`  — constants + all data types (~200 SLOC)
  - `dashboard/format.rs` — pure formatting + layout helpers (~100 SLOC)
  - `dashboard/render.rs` — ratatui widget construction + frame drawing (~150 SLOC)
  - `dashboard/tests.rs`  — moved test block verbatim (~250 SLOC)
  - `dashboard/mod.rs`    — thin re-export facade (~25 SLOC)

Zero behaviour change; all 219 trusty-common tests pass with --features monitor-tui.
Removes the grandfathered dashboard.rs entry from .line-cap-allowlist.tsv.
